backend/requirements.txt
# requirements.txt
# Python依赖包列表

Flask==2.3.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0